rename soundxyz to soundxtz-call-tokenuri
