+++ test/lifecycle_test.mjs
@@ -62,7 +62,9 @@ test("reading a file by line using the line reader", async (t) => {
   };
   const strategies = (
     await loadStrategies("./strategies", "transformer.mjs")
-  ).filter((strategy) => strategy && strategy.name === "soundxyz");
+  ).filter(
+    (strategy) => strategy && strategy.name === "soundxyz-call-tokenuri"
+  );
   const strategy = { ...strategies[0].module, onLine: lineHandlerMock };
   await lineReader(path, strategy);
   t.is(count, 2);
@@ -72,7 +74,9 @@ test("applying transformation strategies to a file", async (t) => {
   const dataPath = resolve(__dirname, "./fixtures/file1.data");
   const strategies = (
     await loadStrategies("./strategies", "transformer.mjs")
-  ).filter((strategy) => strategy && strategy.name === "soundxyz");
+  ).filter(
+    (strategy) => strategy && strategy.name === "soundxyz-call-tokenuri"
+  );
   await lineReader(dataPath, strategies[0].module);
   t.pass();
 });
