+++ test/strategies/soundxyz-call-tokenuri/extractor_input_data
