+++ test/strategies/soundxyz-call-tokenuri/extractor_snapshot.json
